Added extended LBA support and other modifications.
    The major change is the support added for extended LBA. All the
    tests now allocate separate meta buffers only when the namspc
    is of configured for separate meta buffer. When namspc is configured
    for extended LBA format then the allocated buffer is resized to fit
    the meta data as an addendum to data buffer.

    The following are the additional modifications:

    1) Defined maximum data transfer size in tnvme.h and
       used in all places needed.
    2) In metaData.h fixed to return 32 bits from GetMetaBufferSize()
       instead of 16 bits.
    3) Bug fixes in GrpBasicInit/VerifyDataPat and WriteDataPat to
       consider the type of namespace and necessary meta data support
       before sending down any NVM cmds.
    4) Added new test GrpNVMWriteCmd/LbaOutOfRangeMeta_r10b.
    5) Peppered code with LOG_NRM prints to aid in debug sessions.

+++ Cmds/metaData.h
@@ -68,7 +68,7 @@ public:
      *      data buffer is in use/allocated.
      */
     uint8_t *GetMetaBuffer() { return mMetaData.buf; }
-    uint16_t GetMetaBufferSize() const { return mMetaData.size; }
+    uint32_t GetMetaBufferSize() const { return mMetaData.size; }
     uint32_t GetMetaBufferID() const { return mMetaData.ID; }
 
     /// Solely used to notify dnvme how to issue a PRP data buffer
